Add an index to SkGlyphPos
In the mask case, the original position is needed for the fallback case.
Add an index from the original sequence to lookup the source position
when needed.

+++ src/core/SkStrikeCache.cpp
@@ -37,7 +37,7 @@ public:
     }
 
     int glyphMetrics(
-            const SkGlyphID id[], const SkPoint point[], int n, SkGlyphPos result[]) override {
+            const SkGlyphID id[], const SkPoint point[], size_t n, SkGlyphPos result[]) override {
         return fStrike.glyphMetrics(id, point, n, result);
     }
